Wrap TreeCounter app with AuthProvider, refactor TreeCounter app, factor out Routes
