+++ app/config/index.js.dist
@@ -18,5 +18,8 @@ export const context = {
   debug: true, // local console debugging switch
   currency: 'USD',
   mapIds: { inventory: 'dee6acf9de774fe6878813f707b4ab88'},
-  bugsnagApiKey: ''
+  bugsnagApiKey: '',
+  auth0domain: 'https://accounts.plant-for-the-planet.org',
+  auth0clientId: ''
+}
 };
